COMMIT:
Fixed previous_offline system - previous behaviour was that channels defaulted to False causing channels that were offline when the tracker started, to not update the first time they come online. new behaviour is when we check if channel is offline at the start of the tick loop we set previous_offline correctly at the end of the check.